# specs/001-campaign-management/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tome API — Campaign & Character Management
  version: 0.1.0
  description: >
    BFF API for Tome v1 (feature 001-campaign-management). All endpoints require an authenticated
    session (Hive OAuth via the BFF; session cookie). Authorization is enforced server-side on every
    read/write and mirrored on the SSE stream. Sheet `data` is a **typed, discriminated union** keyed
    by `ruleSetId` (ADR-001): each rule set has its own typed sheet schema (see `Sheet` / `DnD35Sheet`),
    the codegen source for the web's sheet types. Derived values are `readOnly` — present in responses
    (computed on read), ignored on write, never stored. Mutable aggregates use optimistic concurrency:
    send the `version` you read; a stale write returns 409.
servers:
  - url: /api
security:
  - sessionCookie: []
paths:
...
  /characters:
    get:
      summary: List the caller's characters
      responses:
        "200":
          description: Characters owned by the caller
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CharacterSummary" }
    post:
      summary: Create a character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, data]
              properties:
                name: { type: string, description: "promoted top-level name for lists/rosters" }
                data:
                  allOf: [ { $ref: "#/components/schemas/SheetInput" } ]
                  description: >-
                    The typed base sheet. Its `ruleSetId` selects the rule set (no separate top-level
                    `ruleSetId` — ADR-001); an unknown/unsupported value is rejected → 400. Partial is fine.
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Character" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /characters/{characterId}:
    parameters: [ { $ref: "#/components/parameters/CharacterId" } ]
    get:
      summary: Get a character (owner, or DM of a campaign it belongs to)
      responses:
        "200":
          description: Character with computed derived values and any soft warnings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Character" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      summary: Update a character sheet (optimistic concurrency)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version, data]
              properties:
                name: { type: string }
                data:
                  allOf: [ { $ref: "#/components/schemas/SheetInput" } ]
                  description: "The typed base sheet. `data.ruleSetId` must match the character's (rule set is fixed for life, FR-002) or the service answers 400."
                version: { type: integer }
      responses:
        "200":
          description: Updated; body includes recomputed derived values and soft warnings (never blocks)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Character" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      summary: Delete a character (owner only; must not be in a campaign)
      responses:
        "204": { description: Deleted }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /campaigns/{campaignId}/rolls:
    parameters: [ { $ref: "#/components/parameters/CampaignId" } ]
    post:
      summary: Make a campaign-scoped roll (appended to campaign.rolls[])
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/RollRequest" } } } }
      responses:
        "201": { description: Recorded roll (also fanned out over SSE), content: { application/json: { schema: { $ref: "#/components/schemas/Roll" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /campaigns/{campaignId}/sessions/{sessionId}/rolls:
    parameters:
      - { $ref: "#/components/parameters/CampaignId" }
      - { $ref: "#/components/parameters/SessionId" }
    post:
      summary: Make a session-scoped roll (appended to session.rolls[])
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/RollRequest" } } } }
      responses:
        "201": { description: Recorded roll (also fanned out over SSE), content: { application/json: { schema: { $ref: "#/components/schemas/Roll" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /campaigns/{campaignId}/sessions/{sessionId}/encounters/{encounterId}/rolls:
    parameters:
      - { $ref: "#/components/parameters/CampaignId" }
      - { $ref: "#/components/parameters/SessionId" }
      - { $ref: "#/components/parameters/EncounterId" }
    post:
      summary: Make an encounter-scoped roll (appended to encounter.rolls[])
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/RollRequest" } } } }
      responses:
        "201": { description: Recorded roll (also fanned out over SSE), content: { application/json: { schema: { $ref: "#/components/schemas/Roll" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: TOME_SESSION
  parameters:
...
    CharacterId: { name: characterId, in: path, required: true, schema: { type: string } }
    CampaignId: { name: campaignId, in: path, required: true, schema: { type: string } }
    SessionId: { name: sessionId, in: path, required: true, schema: { type: string } }
    EncounterId: { name: encounterId, in: path, required: true, schema: { type: string } }
  responses:
...
    Forbidden: { description: Authenticated but not authorized, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
    NotFound: { description: Not found, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
    BadRequest: { description: Invalid request, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
    Conflict: { description: Stale version or constraint conflict (optimistic concurrency), content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
  schemas:
    Problem:
      type: object
      description: RFC-7807 problem detail
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
    RuleWarning:
      type: object
      properties:
        code: { type: string }
        field: { type: string, nullable: true, description: "null for a sheet-wide warning" }
        message: { type: string }
...
    SheetInput:
      oneOf:
        - $ref: "#/components/schemas/DnD35SheetInput"
        - $ref: "#/components/schemas/DarkSoulsSheetInput"
      discriminator:
        propertyName: ruleSetId
        mapping:
          dnd35: "#/components/schemas/DnD35SheetInput"
          darksouls: "#/components/schemas/DarkSoulsSheetInput"
      description: >-
        The base inputs for one rule set — what create/update send and storage keeps; no derived values.
        `ruleSetId` selects the variant; an unknown value is rejected → 400.
    DnD35SheetInput:
      type: object
      description: "D&D 3.5 base inputs (mirrors DnD35CharacterBaseData). Every field optional (sensible defaults)."
      properties:
        ruleSetId: { const: dnd35, description: "discriminator" }
        name: { type: string }
        player: { type: string }
        race: { type: string }
        characterClass: { type: string }
        alignment: { type: string }
        deity: { type: string }
        size: { type: string }
        level: { type: integer }
        experience: { type: integer }
        abilities: { $ref: "#/components/schemas/DnD35AbilityScores" }
        hitPoints: { $ref: "#/components/schemas/DnD35HitPoints" }
        defense: { $ref: "#/components/schemas/DnD35DefenseInputs" }
        saves: { $ref: "#/components/schemas/DnD35SaveInputs" }
        baseAttackBonus: { type: integer }
        grappleSizeMod: { type: integer, description: "grapple special-size scale (Small −4, Large +4)" }
        attacks: { type: array, items: { $ref: "#/components/schemas/DnD35AttackRowInput" } }
        skills: { type: array, items: { $ref: "#/components/schemas/DnD35SkillRowInput" } }
        feats: { type: array, items: { $ref: "#/components/schemas/DnD35FeatRow" } }
        gear: { type: array, items: { $ref: "#/components/schemas/DnD35GearRow" } }
        languages: { type: array, items: { type: string } }
        notes: { type: string }
        spellcasting: { $ref: "#/components/schemas/DnD35SpellcastingInputs" }
    DarkSoulsSheetInput:
      type: object
      description: "Stub base inputs (US5) — present so the union is total."
      properties:
        ruleSetId: { const: darksouls, description: "discriminator" }
        name: { type: string }

    # ---- base value objects (inputs) ----
    DnD35AbilityScores:
      type: object
      properties:
        strength: { type: integer }
        dexterity: { type: integer }
        constitution: { type: integer }
        intelligence: { type: integer }
        wisdom: { type: integer }
        charisma: { type: integer }
    DnD35HitPoints:
      type: object
      description: "Entered hit points — both base inputs (HP is entered in 3.5, not derived). Shared by input and enriched."
      properties:
        max: { type: integer }
        current: { type: integer }
    DnD35DefenseInputs:
      type: object
      properties:
        armorBonus: { type: integer }
        shieldBonus: { type: integer }
        naturalArmor: { type: integer }
        deflection: { type: integer }
        dodge: { type: integer }
        sizeMod: { type: integer, description: "AC/attack size scale (Small +1, Large −1)" }
    DnD35SaveInputs:
      type: object
      properties:
        fortBase: { type: integer }
        refBase: { type: integer }
        willBase: { type: integer }
    DnD35SpellcastingInputs:
      type: object
      properties:
        casterClass: { type: string }
        casterLevel: { type: integer }
        spellKeyAbility: { type: string, description: "select: an ability-mod id, e.g. intMod" }
        spellSlots: { type: array, items: { $ref: "#/components/schemas/DnD35SpellSlotRowInput" } }
        spells: { type: array, items: { $ref: "#/components/schemas/DnD35SpellRow" } }
    DnD35AttackRowInput:
      type: object
      properties:
        weapon: { type: string }
        ability: { type: string, description: "select: ability-mod id (e.g. strMod)" }
        misc: { type: integer }
        damage: { type: string }
        critical: { type: string }
        range: { type: string }
        notes: { type: string }
    DnD35SkillRowInput:
      type: object
      properties:
        skill: { type: string }
        keyAbility: { type: string, description: "select: ability-mod id" }
        ranks: { type: integer }
        classSkill: { type: boolean }
        misc: { type: integer }
    DnD35SpellSlotRowInput:
      type: object
      properties:
        spellLevel: { type: integer }
        slotsPerDay: { type: integer }
        known: { type: integer }
        prepared: { type: integer }

    # ---- response / enriched (served) ----
    Sheet:
      oneOf:
        - $ref: "#/components/schemas/DnD35Sheet"
        - $ref: "#/components/schemas/DarkSoulsSheet"
      discriminator:
        propertyName: ruleSetId
        mapping:
          dnd35: "#/components/schemas/DnD35Sheet"
          darksouls: "#/components/schemas/DarkSoulsSheet"
      description: >-
        The enriched sheet a response returns: base inputs plus derived values (grouped, `readOnly`,
        computed on read, never persisted). Built from the stored SheetInput by `enrich()`.
    DnD35Sheet:
      type: object
      description: "D&D 3.5 enriched sheet (mirrors DnD35CharacterData). `readOnly` properties are derived."
      properties:
        ruleSetId: { const: dnd35, description: "discriminator" }
        name: { type: string }
        player: { type: string }
        race: { type: string }
        characterClass: { type: string }
        alignment: { type: string }
        deity: { type: string }
        size: { type: string }
        level: { type: integer }
        experience: { type: integer }
        abilities: { $ref: "#/components/schemas/DnD35Abilities" }
        hitPoints: { $ref: "#/components/schemas/DnD35HitPoints" }
        defense: { $ref: "#/components/schemas/DnD35Defense" }
        saves: { $ref: "#/components/schemas/DnD35Saves" }
        baseAttackBonus: { type: integer }
        grappleSizeMod: { type: integer }
        initiative: { type: integer, readOnly: true }
        grapple: { type: integer, readOnly: true }
        attacks: { type: array, items: { $ref: "#/components/schemas/DnD35AttackRow" } }
        skills: { type: array, items: { $ref: "#/components/schemas/DnD35SkillRow" } }
        feats: { type: array, items: { $ref: "#/components/schemas/DnD35FeatRow" } }
        gear: { type: array, items: { $ref: "#/components/schemas/DnD35GearRow" } }
        totalWeight: { type: integer, readOnly: true, description: "sum of gear weight" }
        languages: { type: array, items: { type: string } }
        notes: { type: string }
        spellcasting: { $ref: "#/components/schemas/DnD35Spellcasting" }
    DarkSoulsSheet:
      type: object
      description: "Stub enriched variant (US5)."
      properties:
        ruleSetId: { const: darksouls, description: "discriminator" }
        name: { type: string }

    # ---- enriched value objects (base + derived) ----
    DnD35Abilities:
      type: object
      description: "Ability scores plus derived modifiers."
      properties:
        strength: { type: integer }
        dexterity: { type: integer }
        constitution: { type: integer }
        intelligence: { type: integer }
        wisdom: { type: integer }
        charisma: { type: integer }
        strMod: { type: integer, readOnly: true }
        dexMod: { type: integer, readOnly: true }
        conMod: { type: integer, readOnly: true }
        intMod: { type: integer, readOnly: true }
        wisMod: { type: integer, readOnly: true }
        chaMod: { type: integer, readOnly: true }
    DnD35Defense:
      type: object
      description: "Defense bonuses plus derived AC totals."
      properties:
        armorBonus: { type: integer }
        shieldBonus: { type: integer }
        naturalArmor: { type: integer }
        deflection: { type: integer }
        dodge: { type: integer }
        sizeMod: { type: integer }
        armorClass: { type: integer, readOnly: true }
        touchAC: { type: integer, readOnly: true }
        flatFootedAC: { type: integer, readOnly: true }
    DnD35Saves:
      type: object
      description: "Save bases plus derived totals (base + governing ability mod)."
      properties:
        fortBase: { type: integer }
        refBase: { type: integer }
        willBase: { type: integer }
        fortitude: { type: integer, readOnly: true }
        reflex: { type: integer, readOnly: true }
        will: { type: integer, readOnly: true }
    DnD35Spellcasting:
      type: object
      description: "Spellcasting inputs plus derived save-DC base and per-slot totals."
      properties:
        casterClass: { type: string }
        casterLevel: { type: integer }
        spellKeyAbility: { type: string }
        saveDcBase: { type: integer, readOnly: true, description: "10 + key-ability mod; a level-N spell's DC = base + N" }
        spellSlots: { type: array, items: { $ref: "#/components/schemas/DnD35SpellSlotRow" } }
        spells: { type: array, items: { $ref: "#/components/schemas/DnD35SpellRow" } }
    DnD35AttackRow:
      type: object
      properties:
        weapon: { type: string }
        ability: { type: string, description: "select: ability-mod id (e.g. strMod)" }
        misc: { type: integer }
        damage: { type: string }
        critical: { type: string }
        range: { type: string }
        notes: { type: string }
        attackBonus: { type: integer, readOnly: true, description: "baseAttackBonus + ability mod + misc" }
    DnD35SkillRow:
      type: object
      properties:
        skill: { type: string }
        keyAbility: { type: string }
        ranks: { type: integer }
        classSkill: { type: boolean }
        misc: { type: integer }
        total: { type: integer, readOnly: true, description: "ranks + key-ability mod + misc" }
    DnD35SpellSlotRow:
      type: object
      description: "One row per spell level 0–9. slotsPerDay entered; bonus/total derived."
      properties:
        spellLevel: { type: integer }
        slotsPerDay: { type: integer }
        known: { type: integer }
        prepared: { type: integer }
        bonusSpells: { type: integer, readOnly: true }
        total: { type: integer, readOnly: true, description: "slotsPerDay + bonusSpells" }

    # ---- shared rows (no derived; same in input and enriched) ----
    DnD35FeatRow:
      type: object
      properties:
        name: { type: string }
        type: { type: string, enum: [general, combat, metamagic, itemCreation] }
        description: { type: string }
    DnD35GearRow:
      type: object
      properties:
        item: { type: string }
        quantity: { type: integer }
        weight: { type: integer }
        notes: { type: string }
    DnD35SpellRow:
      type: object
      properties:
        spell: { type: string, description: "spell id from the class-filtered spells catalog" }
        level: { type: integer, description: "spell level for the caster class (catalog-supplied)" }
        prepared: { type: integer }
        notes: { type: string }
    CharacterSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        ruleSetId: { type: string }
    Character:
      allOf:
        - $ref: "#/components/schemas/CharacterSummary"
        - type: object
          properties:
            ownerId: { type: string }
            data: { $ref: "#/components/schemas/Sheet" }
            warnings: { type: array, items: { $ref: "#/components/schemas/RuleWarning" } }
            version: { type: integer }
...
    Roll:
      type: object
      description: "Embedded in its container's rolls[] (campaign/session/encounter); scope is the container, so no campaignId/sessionId/encounterId fields"
      properties:
        id: { type: string }
        initiatorId: { type: string, description: "Hive subject who initiated the roll" }
        expression: { type: string }
        results: { type: array, items: { type: integer } }
        total: { type: integer }
        appliedTo:
          type: object
          nullable: true
          properties:
            targetType: { type: string }
            targetId: { type: string }
            field: { type: string }
    RollRequest:
      type: object
      required: [expression]
      description: "Roll-creation body; the container (campaign/session/encounter) is given by the URL path"
      properties:
        expression: { type: string, examples: ["2d6+3", "4d6kh3", "1d20+5"] }
        applyTo:
          type: object
          description: Optional target; applied through the normal sheet-write path (soft warnings apply)
          properties:
            targetType: { type: string, enum: [character, npc] }
            targetId: { type: string }
            field: { type: string, examples: ["hp_current"] }
...